Moves SPI classes under org.partiql.spi
