+++ partiql-eval/src/main/kotlin/org/partiql/eval/internal/operator/rex/ExprSubqueryRow.kt
@@ -1,11 +1,11 @@
 package org.partiql.eval.internal.operator.rex
 
-import org.partiql.errors.CardinalityViolation
 import org.partiql.eval.Environment
 import org.partiql.eval.ExprRelation
 import org.partiql.eval.ExprValue
 import org.partiql.eval.internal.helpers.IteratorSupplier
 import org.partiql.eval.internal.helpers.ValueUtility.check
+import org.partiql.spi.errors.CardinalityViolation
 import org.partiql.spi.value.Datum
 import org.partiql.types.PType
