+++ partiql-spi/src/main/java/org/partiql/spi/value/Datum.java
@@ -5,8 +5,8 @@ import com.amazon.ionelement.api.ElementLoader;
 import com.amazon.ionelement.api.IonElementLoader;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-import org.partiql.errors.DataException;
-import org.partiql.spi.value.ion.IonVariant;
+import org.partiql.spi.errors.DataException;
+import org.partiql.spi.internal.value.ion.IonVariant;
 import org.partiql.types.PType;
 import java.math.BigDecimal;
 import java.math.BigInteger;
