+++ partiql-spi/src/main/kotlin/org/partiql/spi/function/builtins/FnDateAddMonth.kt
@@ -3,8 +3,8 @@
 
 package org.partiql.spi.function.builtins
 
-import org.partiql.errors.DataException
-import org.partiql.errors.TypeCheckException
+import org.partiql.spi.errors.DataException
+import org.partiql.spi.errors.TypeCheckException
 import org.partiql.spi.function.Parameter
 import org.partiql.spi.function.utils.FunctionUtils
 import org.partiql.spi.value.Datum
