+++ partiql-spi/src/main/kotlin/org/partiql/spi/function/builtins/FnIsChar.kt
@@ -3,7 +3,7 @@
 
 package org.partiql.spi.function.builtins
 
-import org.partiql.errors.TypeCheckException
+import org.partiql.spi.errors.TypeCheckException
 import org.partiql.spi.function.Parameter
 import org.partiql.spi.function.utils.FunctionUtils
 import org.partiql.spi.value.Datum
